Add files via upload
main_DPN_v0.py is the original main function with a few extra lines to visualize the data. 
functionSimplex.py- The Dirichlet distribution constructed from the exponential of the DPN output is visualized with a simplex.
The classification accuracy is 100% but the problem I am facing is that the alpha = exp(output) is less than 10 for both indistribution data even though my target distribution has is a dirichilet with precision close to 100. 
In test_testing.py I have a (0,0)as a fixed test point and I keep changing my indistribution area to observe the effect on \alpha